+++ Features/Paths/AssassinsPath.cs
@@ -119,7 +119,7 @@ public static class AssassinsPath
                 "and can expend psionic focus for +7 bonus damage on your next strike (maneuver)."))
             .SetIcon(icon)
             .SetIsClassFeature()
-            .AddFacts([trance.ToString()])
+            .AddFeatureOnClassLevel(feature: trance.ToString(), level: 3, clazz: Guids.PsychicWarriorClass)
             .AddFeatureIfHasFact(checkedFact: Guids.MartialPowerFeature, feature: Guids.MartialPowerAssassinsManeuver)
             .AddPrerequisiteNoFeature(Guids.AssassinsPath)
             .Configure();