+++ Features/Paths/FeralWarriorPath.cs
@@ -109,7 +109,7 @@ public static class FeralWarriorPath
                 "You focus on natural weapon and unarmed combat. You gain a +1 competence bonus to attack rolls (trance) and can expend psionic focus for +7 bonus damage on your next strike (maneuver)."))
             .SetIcon(icon)
             .SetIsClassFeature()
-            .AddFacts([trance.ToString()])
+            .AddFeatureOnClassLevel(feature: trance.ToString(), level: 3, clazz: Guids.PsychicWarriorClass)
             .AddFeatureIfHasFact(checkedFact: Guids.MartialPowerFeature, feature: Guids.MartialPowerFeralWarriorManeuver)
             .AddPrerequisiteNoFeature(Guids.FeralWarriorPath)
             .Configure();